+++ Assets/LDtkUnity/Editor/Utility/RelativeGetter/LDtkRelativeGetter.cs
@@ -15,7 +15,7 @@ namespace LDtkUnity.Editor
         /// <summary>
         /// Will return null with no errors if the load failed unless for a special occasion
         /// </summary>
-        public TAsset GetRelativeAsset(TData def, string relativeTo, LoadAction loadAction = null)
+        public virtual TAsset GetRelativeAsset(TData def, string relativeTo, LoadAction loadAction = null)
         {
             return GetAssetRelativeToAssetPath(relativeTo, GetRelPath(def), loadAction);
         }